Update utilities.py
data_manipulation_sh(network): find names of new components correctly

+++ etrago/tools/utilities.py
@@ -292,10 +292,9 @@ def data_manipulation_sh (network):
     from geoalchemy2.shape import from_shape, to_shape
     
     #add connection from Luebeck to Siems
-
-    new_bus = str(int(network.buses.index.max())+1)
-    new_trafo = str(int(network.transformers.index.max())+1)
-    new_line = str(int(network.lines.index.max())+1)
+    new_bus = str(network.buses.index.astype(np.int64).max()+1)
+    new_trafo = str(network.transformers.index.astype(np.int64).max()+1)
+    new_line = str(network.lines.index.astype(np.int64).max()+1)
     network.add("Bus", new_bus,carrier='AC', v_nom=220, x=10.760835, y=53.909745)
     network.add("Transformer", new_trafo, bus0="25536", bus1=new_bus, x=1.29960, tap_ratio=1, s_nom=1600)
     network.add("Line",new_line, bus0="26387",bus1=new_bus, x=0.0001, s_nom=1600)
